requirements.txt
streamlit
mysql-connector-python
openai
numpy
scikit-learn
pandas
python-dotenv
openpyxl
